+++ django_spire/auth/user/seed.py
@@ -1,5 +1,0 @@
-from django_spire.auth.user.tests.factories import create_super_user
-
-
-print('Creating super user...')
-create_super_user()